done in the god willing

+++ CPP01/ex01/zombieHorde.cpp
@@ -6,7 +6,7 @@
 /*   By: tmiftah <tmiftah@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/20 10:11:37 by tmiftah           #+#    #+#             */
-/*   Updated: 2023/09/09 18:19:59 by tmiftah          ###   ########.fr       */
+/*   Updated: 2023/09/13 12:37:54 by tmiftah          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 Zombie* zombieHorde( int N, std::string name ){
     Zombie *zombie = new Zombie[N];
-    for (size_t i = 0; i < N; i++)
+    for (int i = 0; i < N; i++)
         zombie[i].setZombieName(name);
     return (zombie);
 }
